+++ src/setup_fastsql/__init__.py
@@ -1,3 +1,0 @@
-"""Setup FastSQL - Interactive FastAPI Project Scaffolder"""
-
-__version__ = "0.1.0"